tests: Normalize MODULE_PATH
Use the fs Meson module if we are on Meson 0.54.0 or later, or merely convert
the '\\' (commonly used in Windows paths for directory separators) to '/'.

This way, we can ensure the loaders.cache file is correctly generated for
running the tests on Visual Studio builds.  The test programs will pass on
Visual Studio builds at this point.